requirements.txt
django==1.7.1
django-filebrowser
django-grappelli
django-tinymce
Pillow